Changed the session files to have integer indexes: xxx_0.json instead of xxx_0.0.json Using the 'CullVacMoveCombi' control now. A few changes to try to slightly reduce the checkstyle complaints.

+++ cowtastrophe-war/src/main/java/uk/ac/bioss/cowtastrophe/SimulationHelper.java
@@ -74,8 +74,8 @@ public class SimulationHelper implements Serializable {
      */
     public final String saveSession(final String sessionId, final double time) {
         String sessionDir = new File(parameters.getDirectory(), sessionId).getAbsolutePath();
-        String sessionFile = new File(sessionDir, sessionId + "_" + time + ".ser").getAbsolutePath();
-        String jsonFile = new File(sessionDir, sessionId + "_" + time + ".json").getAbsolutePath();
+        String sessionFile = new File(sessionDir, sessionId + "_" + ((int)time) + ".ser").getAbsolutePath();
+        String jsonFile = new File(sessionDir, sessionId + "_" + ((int)time) + ".json").getAbsolutePath();
 
         try (ObjectOutputStream serFile
                                 = new ObjectOutputStream(new FileOutputStream(sessionFile));
